% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sunrise.R
\name{calculate_sunrise}
\alias{calculate_sunrise}
\title{Function to calculate sunrise and sunset dates and times for a location.}
\usage{
calculate_sunrise(latitude, longitude, start = NA, end = NA,
  tz = "UTC", ...)
}
\arguments{
\item{latitude}{Latitude of a location.}

\item{longitude}{Longitude of a location.}

\item{start}{Start date, if not used, the default is the system's date.}

\item{end}{End date, if not used, the default is the system's date.}

\item{tz}{Time zone to conduct calculations in.}

\item{...}{A construct to allow for absorption of additional arguments, 
useful when using function with \strong{purrr}.}
}
\value{
Tibble.
}
\description{
\code{calculate_sunrise} uses \code{\link{sunriset}} for the calculation of
dates and times.
}
\examples{

# Get sunrise and sunset dates for London for today
calculate_sunrise(latitude = 51.5072, longitude = 0.1275)

# Or specify dates
calculate_sunrise(
  latitude = 51.5072, 
  longitude = 0.1275, 
  start = "2015-12-01", 
  end = "2015-12-31"
)

}
\seealso{
\code{\link{sunriset}}
}
\author{
Stuart K. Grange.
}
