% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_to_spatial.R
\name{data_frame_to_line}
\alias{data_frame_to_line}
\title{Function for converting a data frame into a spatial-lines data frame.}
\usage{
data_frame_to_line(df, latitude = "latitude", longitude = "longitude",
  projection = "+proj=longlat +datum=WGS84 +no_defs", id = NA)
}
\arguments{
\item{df}{Data frame to be converted into SpatialLinesDataFrame.}

\item{latitude}{\code{df}'s latitude variable name.}

\item{longitude}{\code{df}'s longitude variable name.}

\item{projection}{\code{df}'s latitude and longitude projection system. 
Default is WGS84.}

\item{id}{What variable in \code{df} should be used to create separate line}
}
\description{
\code{data_frame_to_line} returns a SpatialLinesDataFrame with a single line 
object. Multiple identifiers are not currently preserved.
}
\details{
In general usage, \code{latitude} and \code{longitude} will be projected in 
WGS84. \code{latitude} and \code{longitude} must not contain \code{NA} values 
because this is a limitation of the spatial object.
}
\examples{
\dontrun{
sp_lines <- data_frame_to_line(data_gps_track, "latitude", "longitude")
}

}
\author{
Stuart K. Grange
}

