% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_to_spatial.R
\name{data_frame_to_lines}
\alias{data_frame_to_lines}
\title{Function for converting a data frame to a SpatialLinesDataFrame.}
\usage{
data_frame_to_lines(df, latitude = "latitude", longitude = "longitude",
  projection = "+proj=longlat +datum=WGS84 +no_defs", id = NA)
}
\arguments{
\item{df}{Data frame to be converted into a SpatialLinesDataFrame.}

\item{latitude}{\code{df}'s latitude variable name.}

\item{longitude}{\code{df}'s longitude variable name.}

\item{projection}{\code{df}'s latitude and longitude projection system. 
Default is WGS84.}

\item{id}{What variable in \code{df} should be used to create separate line}
}
\description{
Function for converting a data frame to a SpatialLinesDataFrame.
}
\examples{
\dontrun{
sp_lines <- data_frame_to_lines(data_gps_track, "latitude", "longitude")
}

}
\author{
Stuart K. Grange
}

