% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_to_spatial.R
\name{data_frame_to_points}
\alias{data_frame_to_points}
\title{Function for converting a data frame into a spatial-points data frame.}
\usage{
data_frame_to_points(df, latitude = "latitude", longitude = "longitude",
  projection = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs")
}
\arguments{
\item{df}{Data frame to be converted into SpatialPointsDataFrame.}

\item{latitude}{\code{df}'s latitude variable name.}

\item{longitude}{\code{df}'s longitude variable name.}

\item{projection}{\code{df}'s latitude and longitude projection system. 
Default is WGS84.}
}
\description{
\code{data_frame_to_points} conveniently transforms a data frame to a 
SpatialPointsDataFrame.
}
\details{
In general usage, \code{latitude} and \code{longitude} will be projected in 
WGS84. \code{latitude} and \code{longitude} must not contain \code{NA} values 
because this is a limitation of the spatial object.
}
\examples{
\dontrun{
sp_points <- data_frame_to_line(data_monitoring_sites, "latitude", "longitude")
}

}
\author{
Stuart K. Grange
}

