% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_to_spatial.R
\name{data_frame_to_polygon}
\alias{data_frame_to_polygon}
\alias{data_frame_to_polygons}
\title{Function for converting a data frame to spatial-polygon data frame.}
\usage{
data_frame_to_polygon(df, latitude = "latitude", longitude = "longitude",
  projection = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs",
  force = TRUE)
}
\arguments{
\item{df}{Data frame to be converted into SpatialPolygonsDataFrame.}

\item{latitude}{\code{df}'s latitude variable name.}

\item{longitude}{\code{df}'s longitude variable name.}

\item{projection}{\code{df}'s latitude and longitude projection system. 
Default is WGS84.}

\item{force}{Force all polygons to have the same group? This is needed to keep 
data and polygons within the same unit and therefore allow for correct binding. 
Default is \code{TRUE}.}
}
\description{
\code{data_frame_to_polygon} conveniently transforms a data frame to a 
SpatialPolygonsDataFrame.
}
\details{
\code{data_frame_to_polygon} will create closed polygons by joining the first
and last observations together in a straight line if the input data frame's 
first and last coordinate pairs to not match. \code{data_frame_to_polygon} 
correctly deals with holes if the variable \code{"hole"} is present in the 
input data frame.

In general usage, \code{latitude} and \code{longitude} will be projected in 
WGS84. \code{latitude} and \code{longitude} must not contain \code{NA} values 
because this is a limitation of the spatial object.
}
\examples{
\dontrun{
# Convert a hand-drawn line from an online application to a polygon
sp_polygon <- data_frame_to_polygon(data_drawn, "latitude", "longitude")
}

}
\author{
Stuart K. Grange
}

