% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_to_spatial.R
\name{data_frame_to_polygons}
\alias{data_frame_to_polygons}
\title{Function for converting a data frame to SpatialPolygonsDataFrame.}
\usage{
data_frame_to_polygons(df, latitude = "latitude", longitude = "longitude",
  projection = "+proj=longlat +datum=WGS84 +no_defs", id = NA)
}
\arguments{
\item{df}{Data frame to be converted into SpatialPolygonsDataFrame.}

\item{latitude}{\code{df}'s latitude variable name.}

\item{longitude}{\code{df}'s longitude variable name.}

\item{projection}{\code{df}'s latitude and longitude projection system. 
Default is WGS84.}

\item{What}{variable in \code{df} should be used to create separate line}
}
\description{
\code{data_frame_to_polygon} will create closed polygons by joining the first
and last observations together in a straight line if the input data frame's 
first and last coordinate pairs to not match.
}
\examples{
\dontrun{
# Convert a hand-drawn line from an online application to a polygon
sp_polygon <- data_frame_to_polygons(data_drawn, "latitude", "longitude")
}

}
\author{
Stuart K. Grange
}

