% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms_to_decimal.R
\name{dms_to_decimal}
\alias{dms_to_decimal}
\title{Function to convert degrees, minutes, and seconds (DMS) to decimal degrees.}
\usage{
dms_to_decimal(degrees, minutes, seconds, convert = FALSE)
}
\arguments{
\item{degrees}{Degrees vector.}

\item{minutes}{Minutes vector.}

\item{seconds}{Seconds vector.}

\item{convert}{Force the input to be numeric vectors. Default is \code{FALSE}.}
}
\value{
Numeric vector representing \code{degrees}, \code{minutes}, and
\code{seconds} as decimal degrees.
}
\description{
Function to convert degrees, minutes, and seconds (DMS) to decimal degrees.
}
\author{
Stuart K. Grange
}

