% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_node_data.R
\name{get_osm_node_data}
\alias{get_osm_node_data}
\title{Function to scrape OpenStreetMap's node XML document for data.}
\usage{
get_osm_node_data(id)
}
\arguments{
\item{id}{OpenStreetMap relation ID.}
}
\value{
Named list with \code{values} and \code{attributes} data frames.
}
\description{
Function to scrape OpenStreetMap's node XML document for data.
}
\author{
Stuart K. Grange
}

