% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_node_data.R
\name{get_osm_relation_data}
\alias{get_osm_relation_data}
\title{Function to scrape OpenStreetMap's relation XML documents for data.}
\usage{
get_osm_relation_data(id, progress = "none")
}
\arguments{
\item{id}{OpenStreetMap relation ID.}

\item{progress}{Type of progress bar to display.}
}
\value{
Named list.
}
\description{
Function to scrape OpenStreetMap's relation XML documents for data.
}
\examples{
\dontrun{

# Get data for Forest of Bowland
get_osm_relation_data(6399292)

}

}
\author{
Stuart K. Grange
}
