% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_node_data.R
\name{get_osm_way_data}
\alias{get_osm_way_data}
\title{Function to scrape OpenStreetMap's way XML documents for data.}
\usage{
get_osm_way_data(id)
}
\arguments{
\item{id}{OpenStreetMap way ID.}
}
\value{
Named list.
}
\description{
Function to scrape OpenStreetMap's way XML documents for data.
}
\examples{
\dontrun{

# Get data for York Minster
get_osm_way_data(28750371)

}

}
\author{
Stuart K. Grange
}

