% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sunrise.R
\name{get_sunrise}
\alias{get_sunrise}
\title{Function to calculate sunrise and sunset dates and times for a location.}
\usage{
get_sunrise(latitude, longitude, start = NA, end = NA)
}
\arguments{
\item{latitude}{Latitude of a location.}

\item{longitude}{Longitude of a location.}

\item{start}{Start date.}

\item{end}{End date.}
}
\value{
Data frame.
}
\description{
\code{get_sunrise} uses \code{\link{sunriset}} for the calculation of dates
and times.
}
\examples{

\dontrun{
# Get sunrise and sunset dates for London for the current year
get_sunrise(latitude = 51.5072, longitude = 0.1275)

# Or specify dates
get_sunrise(latitude = 51.5072, longitude = 0.1275, start = "2015-12-01", 
  end = "2015-12-31")

}

}
\seealso{
\code{\link{sunriset}}
}
\author{
Stuart K. Grange
}
