% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/google_elevation.R
\name{google_elevation}
\alias{google_elevation}
\title{Function to find elevation from latitude and longitude coordinate pairs using
Google Maps API.}
\usage{
google_elevation(latitude, longitude, key)
}
\arguments{
\item{latitude}{Numeric vector of latitude values.}

\item{longitude}{Numeric vector of longitude values.}

\item{key}{Google Maps API key.}
}
\description{
The Google Maps API is not open-source, therefore check the licensing
conditions for usage conditions
(\url{https://developers.google.com/maps/terms}). Users of this function
must use the elevation data to display on an Google Map.
}
\author{
Stuart K. Grange
}

