% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/google_geocode.R
\name{google_geocode}
\alias{google_geocode}
\title{Function to find latitude and longitude pairs from an address-string using
Google Maps.}
\usage{
google_geocode(string, override_limit = TRUE)
}
\arguments{
\item{string}{The address string to transform to latitude and longitude pairs}

\item{override_limit}{Should the function attempt to override the 2500 queries
a day limit. Does not always work.}
}
\description{
The Google Maps API is not open-source, therefore check the licensing
conditions for usage conditions
(\url{https://developers.google.com/maps/terms}).
}
\details{
\code{sp_geocode} is a wrapper for \code{ggmap::geocode} with some simple
enhancements to clean the output and avoid messages to the console.
}
\examples{
\dontrun{
google_geocode("bath abbey")
string     address                                                    latitude longitude
bath abbey bath abbey, bath, bath and north east somerset ba1 1lt, uk 51.38148 -2.358735
}
}
\author{
Stuart K. Grange
}

