% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.sp.R
\name{is.ra}
\alias{is.ra}
\title{Function to test if an R object is a Raster* data-type.}
\usage{
is.ra(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\value{
Logical vector with length of 1.
}
\description{
Function to test if an R object is a Raster* data-type.
}
\author{
Stuart K. Grange
}
