% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/left_join_spatial.R
\name{left_join_spatial}
\alias{left_join_spatial}
\title{Function for a point-in-polygon test.}
\usage{
left_join_spatial(df, latitude = "latitude", longitude = "longitude",
  projection = "+proj=longlat +datum=WGS84", polygons = NA)
}
\arguments{
\item{df}{Data frame containing latitude and longitude variables.}

\item{latitude}{\code{df}'s latitude variable name.}

\item{longitude}{\code{df}'s longitude variable name.}

\item{projection}{\code{df}'s latitude and longitude projection system.
Default is WGS84.}

\item{polygons}{A spatial-polygon data frame to be joined to \code{df}.}
}
\description{
\code{left_join_spatial} tests if a point is within a polygon and joins
data from a spatial object (the data component of the spatial-polygon data
frame) if the match is TRUE. This process is analogous to a SQL left join
with the match being a spatial intersection.
}
\details{
Points to be tested are generally stored in a data frame with latitude and
longitude pairs, while the polygons are stored in a spatial-polygon object.
The result is the input data frame with the joined data contained within the
spatial-polygon data frame. Observations which do not match are filled with
\code{NA}.

\code{sp::over} is used for the point-in-polygon test.
}
\examples{
\dontrun{

# Join air quality zones to latitude and longitude pairs
data.join <- left_join_spatial(data.coordinates, latitude = "latitude",
  longitude = "longitude", polygons = shape.file.pm10)


# I am in London, I know my latitude and longitude. Which borough am I in?

# Load shape file containing the London boroughs
shape.file <- sp_read("data_objects/london_boroughs/london_sport.shp")

# Make latitude and longitude a data frame
data.point <- data.frame(latitude = 51.523595, longitude = -0.027114)

# Test point with 33 polygons (boroughs) in the shape file
left_join_spatial(data.point, polygons = shape.file)
"Tower Hamlets"
}
}
\author{
Stuart K. Grange
}
\seealso{
See \code{\link{spTransform}}, \code{\link{over}},
\code{\link{merge}}, \code{\link{join}}
}

