% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_sp.R
\name{map_to_sp}
\alias{map_to_sp}
\title{Function to convert the map database object to spatial polygons.}
\usage{
map_to_sp()
}
\value{
SpatialPolygonsDataFrame.
}
\description{
Function to convert the map database object to spatial polygons.
}
\examples{

# Import the mapping data as spatial polyogons
sp_map <- map_to_sp()

}
\author{
Stuart K. Grange
}
