% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_coordinates.R
\name{pad_coordinates}
\alias{pad_coordinates}
\title{Function to pad a data frame/tibble's coordinates.}
\usage{
pad_coordinates(df, resolution = 1)
}
\arguments{
\item{df}{Data frame/tibble with \code{latitude} and \code{longitude} 
variables.}

\item{resolution}{Resolution to pad coordinates to in metres.}
}
\value{
Tibble.
}
\description{
Function to pad a data frame/tibble's coordinates.
}
\author{
Stuart K. Grange
}
