% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_wgs84.R
\name{projection_wgs84}
\alias{projection_bng}
\alias{projection_google}
\alias{projection_nztm}
\alias{projection_wgs84}
\title{Functions to wrap common projection (PROJ.4) strings.}
\usage{
projection_wgs84()

projection_google()

projection_bng()

projection_nztm()
}
\description{
Functions to wrap common projection (PROJ.4) strings.
}
\author{
Stuart K.Grange
}

