% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_align.R
\name{ra_align}
\alias{ra_align}
\title{Function to align a raster object with another raster object -- useful when
the raster objects' origins are not identical.}
\usage{
ra_align(ra, ra_template)
}
\arguments{
\item{ra}{A raster object to be aligned.}

\item{ra_template}{A raster object used as the template for realignment.}
}
\value{
Raster layer.
}
\description{
Function to align a raster object with another raster object -- useful when
the raster objects' origins are not identical.
}
\seealso{
\code{\link{ra_bind}}
}
\author{
Stuart K. Grange
}
