% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_cell_number.R
\name{ra_cell_number}
\alias{ra_cell_number}
\title{Function to extract a raster object's cell number for spatial points.}
\usage{
ra_cell_number(ra, sp)
}
\arguments{
\item{ra}{A raster object.}

\item{sp}{A \code{SpatialPointsDataFrame} object.}
}
\value{
\code{sp} with an extra \code{cell_raster} variable in the data slot.
}
\description{
Function to extract a raster object's cell number for spatial points.
}
\author{
Stuart K. Grange
}
