% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_disaggregate.R
\name{ra_disaggregate}
\alias{ra_disaggregate}
\title{Function to create a higher resolution ("disaggregate") raster layer.}
\usage{
ra_disaggregate(ra, factor, method = "bilinear")
}
\arguments{
\item{ra}{Raster object.}

\item{factor}{Amount of disaggregation in number of cells.}

\item{method}{Method to use for interpolation.}
}
\value{
Raster object.
}
\description{
Function to create a higher resolution ("disaggregate") raster layer.
}
\examples{
\dontrun{

ra_countries_smooth <- ra_disaggregate(ra_countries, factor = 5)

}

}
\author{
Stuart K. Grange
}
