% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_drill.R
\name{ra_drill}
\alias{ra_drill}
\title{Function to `drill' into a raster object and return values.}
\usage{
ra_drill(ra, sp, method = "simple", fun = mean, na.rm = TRUE)
}
\arguments{
\item{ra}{Raster object.}

\item{sp}{Spatial object to drill into ra to get values for.}

\item{method}{Method for extraction, \code{"simple"} or \code{"bilinear"} for
interpolation.}

\item{fun}{Function to summarise the values.}

\item{na.rm}{Should \code{NA}s be ommited when using \code{fun}?}
}
\value{
Tibble.
}
\description{
\code{ra_drill} wraps the \code{\link{extract}} raster function.
}
\seealso{
\code{\link[raster]{extract}}, \code{\link{tidy_ra_drill}}
}
\author{
Stuart K. Grange
}
