% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_drill_with_cell_number.R
\name{ra_drill_with_cell_number}
\alias{ra_drill_with_cell_number}
\title{Function to `drill' into a raster object and return values based on cell 
numbers.}
\usage{
ra_drill_with_cell_number(ra, cell)
}
\arguments{
\item{ra}{Raster object.}

\item{cell}{A vector of cell numbersto extract values for.}
}
\value{
Tibble.
}
\description{
Function to `drill' into a raster object and return values based on cell 
numbers.
}
\seealso{
\code{\link[raster]{extract}}, \code{\link{ra_drill}}
}
\author{
Stuart K. Grange.
}
