% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_extent.R
\name{ra_extent}
\alias{ra_extent}
\title{Function to create a raster-extent object based on a spatial polygon.}
\usage{
ra_extent(sp)
}
\arguments{
\item{sp}{A spatial polygon.}
}
\value{
A raster-extent object
}
\description{
Function to create a raster-extent object based on a spatial polygon.
}
\author{
Stuart K. Grange
}
