% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_focal.R
\name{ra_focal}
\alias{ra_focal}
\title{Function to calculate focal values for a raster object using a matrix of 
weights.}
\usage{
ra_focal(ra, weights, fun = mean, na.rm = FALSE, na.only = FALSE)
}
\arguments{
\item{ra}{Raster object.}

\item{weights}{A 3 by 3 matrix of values for the focal values.}

\item{fun}{Function to use to sumamrise the focal values}

\item{na.rm}{Should \code{NA}s be ommited from the summaries?}

\item{na.only}{Should only cells with \code{NA}s be replaced with focal 
values?}
}
\value{
Raster object.
}
\description{
\code{ra_focal} has the effect of smoothing a raster object.
}
\examples{
\dontrun{

ra_focal(
  ra_heights, 
  weights = matrix(1, 3, 3), 
  fun = mean, 
  na.rm = TRUE, 
  na.only = TRUE
)

} 

}
\author{
Stuart K. Grange
}
