% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_forify.R
\name{ra_forify}
\alias{ra_forify}
\title{Function to transform a raster object to a data frame (a table).}
\usage{
ra_forify(ra)
}
\arguments{
\item{ra}{Raster object.}
}
\value{
Data frame.
}
\description{
Function to transform a raster object to a data frame (a table).
}
\author{
Stuart K. Grange
}
