% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_from_sp.R
\name{ra_from_sp}
\alias{ra_from_sp}
\title{Function to create a raster object from a spatial (vector) object.}
\usage{
ra_from_sp(sp, resolution, drop_index = TRUE)
}
\arguments{
\item{sp}{Spatial object.}

\item{resolution}{Resolution of the returned raster object.}

\item{drop_index}{Should the index be dropped from the raster return?}
}
\value{
Raster layer or brick.
}
\description{
Function to create a raster object from a spatial (vector) object.
}
\seealso{
\code{\link{rasterize}}
}
\author{
Stuart K. Grange
}
