% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_from_xyz.R
\name{ra_from_xyz}
\alias{ra_from_xyz}
\title{Function to}
\usage{
ra_from_xyz(df, resolution = NA, projection = projection_wgs84(), digits = 5)
}
\arguments{
\item{df}{Data frame containing \code{latitude}, \code{longitude}, and 
\code{value} variables. The points also need to be on a regular grid.}

\item{resolution}{Raster's resolution, in \code{projection}'s units.}

\item{projection}{df's latitude and longitude projection system. Default is 
WGS84.}

\item{digits}{Precision to use for detecting if the points are on a 
rectangular grid.}
}
\value{
Raster layer.
}
\description{
Function to
}
\author{
Stuart K. Grange
}
