% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_get_options.R
\name{ra_get_options}
\alias{ra_get_options}
\alias{ra_set_temporary_directory}
\title{Function to print the \strong{raster} package's options.}
\usage{
ra_get_options()

ra_set_temporary_directory(x)
}
\arguments{
\item{x}{Function specific option.}
}
\description{
Function to print the \strong{raster} package's options.
}
