% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_interpolate.R
\name{ra_interpolate}
\alias{ra_interpolate}
\title{Function to interpolate a raster object.}
\usage{
ra_interpolate(ra, variable = "value", sp,
  method = "inverse_distance_weighted")
}
\arguments{
\item{ra}{Raster object to be interpolated.}

\item{variable}{Variable name to interpolate in \code{ra}.}

\item{sp}{Spatial point object which contains the locations of the raster 
points.}

\item{method}{Interpolation method to use. Only 
\code{inverse_distance_weighted} is implemented.}
}
\value{
Raster layer.
}
\description{
Function to interpolate a raster object.
}
\author{
Stuart K. Grange
}
