% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_intersect_test.R
\name{ra_intersect_test}
\alias{ra_intersect_test}
\title{Function to test if a raster object intersects with another spatial or raster
object.}
\usage{
ra_intersect_test(ra, sp)
}
\arguments{
\item{ra}{Raster object.}

\item{sp}{Spatial or raster object.}
}
\value{
Logical vector with length of 1.
}
\description{
Function to test if a raster object intersects with another spatial or raster
object.
}
\author{
Stuart K. Grange
}
