% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_load.R
\name{ra_load}
\alias{ra_load}
\title{Function to explicitly load a raster object into physical memory.}
\usage{
ra_load(ra)
}
\arguments{
\item{ra}{Raster object.}
}
\value{
Raster object.
}
\description{
Function to explicitly load a raster object into physical memory.
}
