% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_raster.R
\name{ra_raster}
\alias{ra_raster}
\alias{ra_set_values}
\title{Function to create a raster object from either a raster-extent or spatial
object.}
\usage{
ra_raster(x, projection = projection_wgs84(), resolution = NA, values = NA)

ra_set_values(ra, values)
}
\arguments{
\item{x}{Raster-extent or spatial object.}

\item{projection}{\code{x}'s latitude and longitude projection system. 
Default is WGS84.}

\item{resolution}{A vector with the length of 1 or 2 indicating the 
resolution of the raster object, in the units of \code{projection}.}

\item{values}{Values for the raster object to have. This vector is passed to
\code{\link{ra_set_values}}.}

\item{ra}{Raster object.}
}
\description{
Function to create a raster object from either a raster-extent or spatial
object.
}
