% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_read.R
\name{ra_read}
\alias{ra_read}
\alias{ra_read_brick}
\title{Function to read a raster object from file.}
\usage{
ra_read(file, variable = NA, in_memory = FALSE)

ra_read_brick(file, variable = NA, level = 1, in_memory = FALSE)
}
\arguments{
\item{file}{File containing raster data.}

\item{variable}{An optional variable to read within \code{file}.}

\item{in_memory}{Should the raster object be forced into physical memory?}

\item{level}{What level to load, usually a fourth dimension variable.}
}
\value{
Raster object.
}
\description{
Function to read a raster object from file.
}
\seealso{
\code{\link[raster]{raster}}
}
\author{
Stuart K. Grange
}
