% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_read_nested.R
\name{ra_read_nested}
\alias{ra_read_nested}
\title{Function to read raster files within a nested tibble.}
\usage{
ra_read_nested(file, forify = FALSE)
}
\arguments{
\item{file}{Vector of file names containing raster data.}

\item{forify}{Should the raster data also be loaded (or converted/fortified) 
into a tabular form?}
}
\value{
Nested tibble grouped by rows.
}
\description{
Function to read raster files within a nested tibble.
}
\author{
Stuart K. Grange
}
