% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_read_variables.R
\name{ra_read_variables}
\alias{ra_read_variables}
\title{Function to read variables (and levels) in a raster data file.}
\usage{
ra_read_variables(file)
}
\arguments{
\item{file}{Vector of file names containing raster data.}
}
\value{
Tibble.
}
\description{
Function to read variables (and levels) in a raster data file.
}
\author{
Stuart K. Grange
}
