% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_resolution.R
\name{ra_resolution}
\alias{ra_resolution}
\title{Function to get the resolution of a raster object.}
\usage{
ra_resolution(ra)
}
\arguments{
\item{ra}{Raster object.}
}
\value{
Tibble.
}
\description{
Function to get the resolution of a raster object.
}
\author{
Stuart K. Grange
}
