% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_summarise.R
\name{ra_summarise}
\alias{ra_summarise}
\title{Function to summarise each raster layer and return results in a tibble.}
\usage{
ra_summarise(
  ra,
  f = mean,
  na.rm = TRUE,
  by_layer = TRUE,
  variable_as_date = FALSE,
  tz = "UTC"
)
}
\arguments{
\item{ra}{Raster object.}

\item{f}{Summary function. \code{mean} is the default.}

\item{na.rm}{Should \code{NA}s be omitted?}

\item{by_layer}{Should each layer be summarised?}

\item{variable_as_date}{Are the variables/names of the \code{\link{ra_drill}}
output dates? If so, the dates will be parsed and the variable renamed.}

\item{tz}{If \code{variable_as_date} is \code{TRUE}, what time zone are the 
dates stored in? This will almost certainly be \code{UTC}.}
}
\value{
Tibble.
}
\description{
Function to summarise each raster layer and return results in a tibble.
}
\author{
Stuart K. Grange
}
