% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_switch.R
\name{ra_switch}
\alias{ra_switch}
\title{Function to switch a raster object's x coordinates from 0 to 360 to -180 and
180 degrees.}
\usage{
ra_switch(ra)
}
\arguments{
\item{ra}{Raster object.}
}
\value{
Raster object.
}
\description{
Function to switch a raster object's x coordinates from 0 to 360 to -180 and
180 degrees.
}
\seealso{
\code{\link[raster]{rotate}}
}
\author{
Stuart K. Grange.
}
