% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_to_data_frame.R
\name{ra_to_data_frame}
\alias{ra_to_data_frame}
\title{Function to coerce a raster object into a data frame.}
\usage{
ra_to_data_frame(ra, rename = FALSE)
}
\arguments{
\item{ra}{Raster object.}

\item{rename}{Should the variables in the returned data frame be altered?}
}
\value{
Data frame.
}
\description{
Function to coerce a raster object into a data frame.
}
\author{
Stuart K. Grange
}
