% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_to_polygons.R
\name{ra_to_polygons}
\alias{ra_to_polygons}
\alias{ra_to_points}
\title{Function to transform a raster object to spatial polygons.}
\usage{
ra_to_polygons(
  ra,
  n = 4,
  na.rm = TRUE,
  digits = 12,
  dissolve = FALSE,
  data = TRUE
)

ra_to_points(ra)
}
\arguments{
\item{ra}{Raster object.}

\item{n}{Number of nodes which make up the polygon (4, 8, or 16).}

\item{na.rm}{Should raster cells with \code{NA}s be omitted?}

\item{digits}{Number of digits for the coordinates.}

\item{dissolve}{Should polygons with the same values be dissolved into 
multi-polygon objects?}

\item{data}{Should the data slot be dropped?}
}
\value{
Spatial polygons (SpatialPolygonsDataFrame or SpatialPolygons).
}
\description{
Function to transform a raster object to spatial polygons.
}
\seealso{
\code{\link{rasterToPolygons}}
}
\author{
Stuart K. Grange
}
