% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_variable_name.R
\name{ra_variable_name}
\alias{ra_variable_name}
\title{Function to extract the variable name from a raster object.}
\usage{
ra_variable_name(ra)
}
\arguments{
\item{ra}{Raster object.}
}
\value{
Character vector with length of 1.
}
\description{
Function to extract the variable name from a raster object.
}
\author{
Stuart K. Grange.
}
