% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gps_visualizer.R
\name{read_gps_visualizer}
\alias{read_gps_visualizer}
\title{Function to read tab delimited text files generated by 
\code{http://www.gpsvisualizer.com/draw} as spatial polygons.}
\usage{
read_gps_visualizer(file, type = "polygon")
}
\arguments{
\item{file}{File name.}

\item{type}{Type of spatial object to generate.}
}
\value{
Spatial polygons.
}
\description{
Function to read tab delimited text files generated by 
\code{http://www.gpsvisualizer.com/draw} as spatial polygons.
}
\author{
Stuart K. Grange
}
