% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_my_tracks.R
\name{read_my_tracks}
\alias{read_my_tracks}
\title{Function to read tabular (\code{.csv}) data generated by Google's GPS 
MyTracks Android app.}
\usage{
read_my_tracks(file, verbose = FALSE)
}
\arguments{
\item{file}{A vector of file names from Google's MyTracks Android app.}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
The MyTracks app was discontinued in 2016.
}
\author{
Stuart K. Grange
}
