% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_points.R
\name{sort_points}
\alias{sort_points}
\title{Function to sort a set of points/coordinates in a clockwise or anti-clockwise
direction.}
\usage{
sort_points(df, y = "latitude", x = "longitude", clockwise = TRUE)
}
\arguments{
\item{df}{Data frame containing y and x points.}

\item{y}{Name of y variable in \code{df}.}

\item{x}{Name of x variable in \code{df}.}

\item{clockwise}{Should the points be arranged in a clockwise direction? If 
\code{FALSE}, the order will be an anti-clockwise direction.}
}
\description{
\code{sort_points} can be useful for creating spatial lines or polygons from 
points.
}
\author{
Stuart K. Grange
}
