% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_add_area.R
\name{sp_add_area}
\alias{sp_add_area}
\title{Function to calculate areas of a spatial object and add the results to the 
data slot.}
\usage{
sp_add_area(sp, round = NA)
}
\arguments{
\item{sp}{A spatial polygons object with a data slot.}

\item{round}{Number of digits to round the calculated area result to.}
}
\value{
Spatial polygons with an additional variable in the data slot.
}
\description{
Function to calculate areas of a spatial object and add the results to the 
data slot.
}
\author{
Stuart K. Grange
}
