% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_add_row_numbers.R
\name{sp_add_row_numbers}
\alias{sp_add_row_numbers}
\title{Function to add a row number variable to a spatial object's data slot.}
\usage{
sp_add_row_numbers(sp, name = "id", zero_based = FALSE)
}
\arguments{
\item{sp}{Spatial object containing a data slot.}

\item{name}{Name of variable to use.}

\item{zero_based}{Should the numbering start at 0 rather than 1?}
}
\description{
Function to add a row number variable to a spatial object's data slot.
}
\author{
Stuart K. Grange
}
