% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp_area.R
\name{sp_area}
\alias{sp_area}
\title{Function to calculate areas of spatial polygons in metres-squared,
kilometres-squared, or hectares.}
\usage{
sp_area(sp, unit = "m", features = TRUE)
}
\arguments{
\item{sp}{Spatial-polygons object.}

\item{unit}{Unit of returned area. Can be "m", "km", or "ha". Default is "m".}

\item{features}{Should the individual areas of the features contained within
\code{sp} be calculated? Default is TRUE.}
}
\description{
If the spatial polygons are projected in a system with a metre unit,
\code{rgeos::gArea} is used to calculate the area of the polygons. If the
projection system is WGS84, a trigonometric function is used
(\code{geosphere::areaPolygon}). Although using a projection system with
metres yields the most accurate results, the \code{geosphere::areaPolygon}
function has been tested and performs very well.
}
\examples{
\dontrun{
# Calculate the area of the City of London, sp is projected in OSGB36
sp_area(sp.city.of.london, unit = "km")
2.903926
}
}
\author{
Stuart K. Grange
}

