% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_boundary.R
\name{sp_boundary}
\alias{sp_boundary}
\title{Function to get boundary of a spatial object.}
\usage{
sp_boundary(sp, features = FALSE)
}
\arguments{
\item{sp}{Spatial object.}

\item{features}{Create boundaries for features within \code{sp}. Default is
\code{FALSE}.}
}
\description{
\code{sp_boundary} often requires a single object to determine a boundary. 
Use \code{\link{sp_unite}} to create a single geometry from many.
}
\author{
Stuart K. Grange
}

