% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_bounding_box.R
\name{sp_bounding_box}
\alias{sp_bounding_box}
\title{Function to extract bounding box from a spatial object.}
\usage{
sp_bounding_box(sp, vector = TRUE)
}
\arguments{
\item{sp}{Spatial object.}

\item{vector}{Should the bounding box be returned as a vector with the order
xmin, xmax, ymin, ymax? Default is \code{TRUE}.}
}
\description{
Function to extract bounding box from a spatial object.
}
\author{
Stuart K. Grange
}

