% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_buffer.R
\name{sp_buffer}
\alias{sp_buffer}
\title{Function to include or exclude an area of a spatial object.}
\usage{
sp_buffer(sp, features = TRUE, width)
}
\arguments{
\item{sp}{Spatial object.}

\item{features}{Should the geometries of all features within \code{sp} be
buffered? Default is \code{TRUE}.}

\item{width}{Width of buffer. \code{width}'s units is in \code{sp}'s 
projection system.}
}
\description{
Function to include or exclude an area of a spatial object.
}
\author{
Stuart K. Grange
}

