% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_centroid.R
\name{sp_centroid}
\alias{sp_centroid}
\title{Function to find the centroid of a spatial object.}
\usage{
sp_centroid(sp, features = TRUE)
}
\arguments{
\item{sp}{Spatial object.}

\item{features}{Should the centroids of all features within \code{sp} be
calculated? Default is \code{TRUE}.}
}
\description{
Function to find the centroid of a spatial object.
}
\author{
Stuart K. Grange
}

