% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_class.R
\name{sp_class}
\alias{sp_class}
\title{Function to find class of spatial object as a vector with a length of 1.}
\usage{
sp_class(sp)
}
\arguments{
\item{sp}{Spatial object.}
}
\description{
Function to find class of spatial object as a vector with a length of 1.
}
\author{
Stuart K. Grange
}

