% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp_clip.R
\name{sp_clip}
\alias{sp_clip}
\title{Function to clip spatial objects by a rectangular envelope.}
\usage{
sp_clip(sp, envelope)
}
\arguments{
\item{sp}{Spatial object to be clipped.}

\item{envelope}{A vector containing the extend of the envelope to be used as a
bounding box. The order must be: xmin, xmax, ymin, ymax and should be in the
same projection as \code{sp}.}
}
\description{
Function to clip spatial objects by a rectangular envelope.
}
\author{
Stuart K. Grange
}
\seealso{
\link{`[`}
}

