% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_convex_hull.R
\name{sp_convex_hull}
\alias{sp_convex_hull}
\title{Function to create the smallest convex polygon that contains all input
geometries.}
\usage{
sp_convex_hull(sp, features = FALSE)
}
\arguments{
\item{sp}{Spatial object.}

\item{features}{Should the individual polygons of the features contained 
within \code{sp} be calculated? Default is \code{FALSE}.}
}
\description{
\code{sp_convex_hull} is a simple wrapper for \code{rgeos::gConvexHull}.
}
\examples{
\dontrun{
# Create a boundary-polygon containing all points
sp_boundary_polygon <- sp_convex_hull(sp_points)
}

}
\author{
Stuart K. Grange
}

