% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_promote.R
\name{sp_demote}
\alias{sp_demote}
\title{Function to drop data slots in geometries.}
\usage{
sp_demote(sp)
}
\arguments{
\item{sp}{Spatial object}
}
\description{
\code{sp_demote} will remove the data slot in geometries.
}
\examples{
\dontrun{
# Drop data slot
sp_without_data <- sp_promote(sp_with_data)
}

}
\author{
Stuart K. Grange
}

