% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_disaggregate.R
\name{sp_disaggregate}
\alias{sp_disaggregate}
\title{Function to disaggregate spatial objects.}
\usage{
sp_disaggregate(sp)
}
\arguments{
\item{sp}{Spatial object}
}
\description{
\code{sp_disaggregate} will break apart multi-polygons to create many 
polygons and is a simple wrapper for \code{\link{sp::disaggregate}}.
}
\seealso{
\code{\link{sp_unite}}
}
\author{
Stuart K. Grange
}
