% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_dissolve_holes.R
\name{sp_dissolve_holes}
\alias{sp_dissolve_holes}
\title{Function to disolve/drop holes in a spatial polygon object.}
\usage{
sp_dissolve_holes(sp)
}
\description{
\code{sp_dissolve_holes} is useful when polygons contain holes but only the 
outer boundary is desired.
}
\details{
If the individual polygons have data associated with them, it will be 
dropped when the holes are dissolved.
}
\author{
Stuart K. Grange
}
