% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_dissolve_polygons.R
\name{sp_dissolve_polygons}
\alias{sp_dissolve_polygons}
\title{Function to unite polygons together and create a single spatial object, 
\emph{i.e.} disolve.}
\usage{
sp_dissolve_polygons(sp, features = NULL)
}
\arguments{
\item{sp}{Spatial polygons.}

\item{features}{Should features within \code{sp} attempted to be preserved?}
}
\value{
Spatial polygons.
}
\description{
If the individual polygons have data associated with them, it will be 
dropped when the polygons are united.
}
\author{
Stuart K. Grange
}
