% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_erase.R
\name{sp_erase}
\alias{sp_erase}
\title{Function to erase pieces of a polygon based on another polygon.}
\usage{
sp_erase(sp_1, sp_2, features = FALSE)
}
\arguments{
\item{sp_1}{Geometry one.}

\item{sp_2}{Geometry two,}

\item{features}{Should the function keep features?}
}
\description{
\code{sp_erase} can be thought as the inverse of \code{sp_punch}.
}
\author{
Stuart K. Grange
}
