% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_feature_ids.R
\name{sp_feature_ids}
\alias{sp_feature_ids}
\title{Function to get ids from spatial objects.}
\usage{
sp_feature_ids(sp)
}
\arguments{
\item{sp}{Spatial object.}
}
\description{
Function to get ids from spatial objects.
}
\author{
Stuart K. Grange
}
