% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_filter.R
\name{sp_filter}
\alias{sp_filter}
\title{Function to filter a spatial object by another spatial object 
(usually a spatial polygon).}
\usage{
sp_filter(sp, sp_boundary)
}
\arguments{
\item{sp}{Spatial object to be filtered.}

\item{sp_boundary}{Usually a spatial polygon to be used as the filtering 
element for \code{sp}.}
}
\value{
Filtered version of \code{sp}.
}
\description{
Both spatial objects must have identical projection systems.
}
\seealso{
\code{\link{sp_crop}}, \code{\link{sp_intersect}}
}
\author{
Stuart K. Grange
}
