% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_flip.R
\name{sp_flip}
\alias{sp_flip}
\title{Function to flip a spatial object.}
\usage{
sp_flip(sp)
}
\arguments{
\item{sp}{Spatial object.}
}
\description{
Function to flip a spatial object.
}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{elide}}
}

