% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_force_projection.R
\name{sp_force_projection}
\alias{sp_force_projection}
\title{Function to force a spatial object's projection system without transformation.}
\usage{
sp_force_projection(sp, projection = projection_wgs84())
}
\arguments{
\item{sp}{Spatial object which is to be manipulated.}

\item{projection}{A proj4 string for the projection-transformation. Default 
is the WGS84 string.}
}
\value{
Spatial object, the same type as \code{sp}.
}
\description{
Function to force a spatial object's projection system without transformation.
}
\author{
Stuart K. Grange
}
