% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_from_js_array.R
\name{sp_from_js_array}
\alias{sp_from_js_array}
\title{Function to create spatial objects from JavaScript arrays.}
\usage{
sp_from_js_array(text, type = "points", projection = projection_wgs84())
}
\arguments{
\item{type}{Spatial data type. Default is \code{"points"}.}

\item{file}{File, url, or character vector containing a JavaScript array.}
}
\description{
JavaScript arrays take the form of:
\code{[[54.35221,-0.88518],[54.35237,-0.88544]]}.
}
\author{
Stuart K. Grange
}
