% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_intersect.R
\name{sp_intersect}
\alias{sp_intersect}
\title{Function to intersect/clip/crop/filter spatial objects by a spatial polygon.}
\usage{
sp_intersect(sp, sp_polygon)
}
\arguments{
\item{sp}{Spatial object to be clipped.}

\item{sp_polygon}{Spatial polygon used as the filter.}
}
\value{
Spatial data class, the same as \code{sp}.
}
\description{
Function to intersect/clip/crop/filter spatial objects by a spatial polygon.
}
\seealso{
\code{\link{sp_filter}}, \code{\link{sp_clip}}
}
\author{
Stuart K. Grange
}
