% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_is_valid.R
\name{sp_is_valid}
\alias{sp_is_valid}
\title{Function to test if a spatial object is valid.}
\usage{
sp_is_valid(sp, features = TRUE)
}
\arguments{
\item{sp}{Spatial object.}

\item{features}{Should all the sub-geometries be tested?}
}
\value{
Logical vector.
}
\description{
Function to test if a spatial object is valid.
}
\author{
Stuart K. Grange
}
