% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_read.R
\name{sp_layer_info}
\alias{sp_layer_info}
\title{Function to return layer information within a spatial data file.}
\usage{
sp_layer_info(file, layer, geom = NULL)
}
\arguments{
\item{file}{Spatial data file.}

\item{layer}{Layer within \code{file} to read.}

\item{geom}{An override for when spatial data files contain more than one 
geometry type in a layer. This is commonly found in GeoJSON and KML files. 
\code{geom} can be one of: \code{"points"}, \code{"lines"}, or 
\code{"polygons"}.}
}
\description{
Function to return layer information within a spatial data file.
}
\examples{
\dontrun{
# Get infomation of the lines geom within a layer in a GeoJSON file
sp_layer_info("../../data/points_and_lines.json", "OGRGeoJSON", 
              geom = "lines")
}

}
\author{
Stuart K. Grange
}

