% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_read.R
\name{sp_list_drivers}
\alias{sp_list_drivers}
\title{Function to list spatial data file drivers on a system.}
\usage{
sp_list_drivers()
}
\description{
A simple wrapper for \code{rgdal::ogrDrivers}. No arguments are used.
}
\examples{
\dontrun{
# List system's drivers
sp_list_drivers()
}

}
\author{
Stuart K. Grange
}

