% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_read.R
\name{sp_list_layers}
\alias{sp_list_layers}
\title{Function to list layers within a spatial data file.}
\usage{
sp_list_layers(file)
}
\arguments{
\item{file}{Spatial data file.}
}
\description{
Function to list layers within a spatial data file.
}
\examples{
\dontrun{
# List layers in GeoJSON file, there will only be one
sp_list_layers("../../data/points_and_lines.json")
}

}
\author{
Stuart K. Grange
}

