% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_make_grid.R
\name{sp_make_grid}
\alias{sp_make_grid}
\title{Function to make a gridded spatial points object.}
\usage{
sp_make_grid(sp, cellsize)
}
\arguments{
\item{sp}{Spatial polygons.}

\item{cellsize}{Resolution of spatial grid in the projection system of 
\code{sp}.}
}
\value{
Spatial points.
}
\description{
Function to make a gridded spatial points object.
}
\author{
Stuart K. Grange
}
