% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_projection.R
\name{sp_projection}
\alias{sp_projection}
\title{Function to return a spatial object's projection system.}
\usage{
sp_projection(sp)
}
\arguments{
\item{sp}{Spatial object}
}
\description{
Function to return a spatial object's projection system.
}
\author{
Stuart K. Grange
}

