% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_punch.R
\name{sp_punch}
\alias{sp_punch}
\title{Function to "punch" (add/insert) a hole in a polygon.}
\usage{
sp_punch(sp_base, sp_hole, features = FALSE)
}
\arguments{
\item{sp_base}{SpatialPolygons which will have a hole punched into it.}

\item{sp_hole}{SpatialPolygons which will be represented as a hole in 
\code{sp_base}.}

\item{features}{Should \code{sp_punch} be applied across all features in
\code{sp_base}?}
}
\description{
Punching could also be called a reverse-clip or an erasing process.
}
\examples{
\dontrun{
# Load coastline
sp_uk <- sp_read("uk_coastline")

# Load a drawn polygon
data_drawn <- read.delim("drawn_polygon.txt")

# Promote to spatial
sp_drawn <- data_frame_to_polygon(data_drawn)

# Punch a hole in the coastline object
sp_uk_punch <- sp_punch(sp_uk, sp_drawn)

# Check
plot(sp_uk_punch)

}

}
\author{
Stuart K. Grange
}

