% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_read_all_geoms.R
\name{sp_read_all_geoms}
\alias{sp_read_all_geoms}
\title{Function to read all geoms in a spatial data file.}
\usage{
sp_read_all_geoms(file, verbose = FALSE)
}
\arguments{
\item{file}{Spatial data file to read.}

\item{verbose}{Should the function give messages?}
}
\value{
Named list containing spatial objects of different data types.
}
\description{
Function to read all geoms in a spatial data file.
}
\seealso{
\code{\link{readOGR}}, \code{\link{sp_list_drivers}}, 
\code{\link{sp_list_layers}}, \code{\link{sp_read}}
}
\author{
Stuart K. Grange
}
