% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_reflect.R
\name{sp_reflect}
\alias{sp_reflect}
\title{Function to relect/reverse a spatial object.}
\usage{
sp_reflect(sp, x = FALSE, y = FALSE)
}
\arguments{
\item{sp}{Spatial object.}

\item{x}{Should \code{sp} be reflected in the \code{x} diemsion?}

\item{y}{Should \code{sp} be reflected in the \code{y} diemsion?}
}
\description{
Function to relect/reverse a spatial object.
}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{elide}}
}

