% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_reset_feature_ids.R
\name{sp_reset_feature_ids}
\alias{sp_reset_feature_ids}
\title{Function to reset spatial object's feature IDs.}
\usage{
sp_reset_feature_ids(sp)
}
\arguments{
\item{sp}{A spatial object. \code{sp} can also be a list containing many 
spatial objects.}
}
\description{
\code{sp_reset_feature_ids} will reset the feature IDs in a spatial object to
sequential integers (beginning at 1), but these will be stored as a characters.
If the spatial object has a data slot, the row names of the data frame will 
also be reset.
}
\examples{
\dontrun{
# Load data
sp_zones <- sp_read("zones.json")

# Filter
sp_zones <- subset(sp_zones, country == "DE")

# Reset feature ids
sp_zones <- sp_reset_feature_ids(sp_zones)

}

}
\author{
Stuart K. Grange
}

