% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_rotate.R
\name{sp_rotate}
\alias{sp_rotate}
\title{Function to rotate a spatial object.}
\usage{
sp_rotate(sp, amount)
}
\arguments{
\item{sp}{Spatial object.}

\item{amount}{Number of degrees to rotate object in degrees in a clock-wise
direction.}
}
\description{
Function to rotate a spatial object.
}
\seealso{
\code{\link{elide}}
}
\author{
Stuart K. Grange
}
