% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_snap_points_to_lines.R
\name{sp_snap_points_to_lines}
\alias{sp_snap_points_to_lines}
\title{Function to "snap" spatial points to the nearest spatial line.}
\usage{
sp_snap_points_to_lines(sp_points, sp_lines, max_distance = NA)
}
\arguments{
\item{sp_points}{Spatial points to be snapped.}

\item{sp_lines}{Spatial lines to snap \code{sp_points} to.}

\item{max_distance}{The maximum distance to avoid snapping points that are 
farther apart.}
}
\value{
Spatial points with altered coordinates; all points will lie on lines
supplied for \code{sp_lines}.
}
\description{
Function to "snap" spatial points to the nearest spatial line.
}
\author{
Stuart K. Grange
}
