% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_split.R
\name{sp_split}
\alias{sp_split}
\title{Function to split a spatial object by feature.}
\usage{
sp_split(sp)
}
\arguments{
\item{sp}{Spatial object.}
}
\value{
Named list containing \emph{n} spatial objects.
}
\description{
Function to split a spatial object by feature.
}
\author{
Stuart K. Grange
}
