% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_tessellation_polygons.R
\name{sp_tessellation_polygons}
\alias{sp_tessellation_polygons}
\title{Function to create tessellation polygons from spatial points.}
\usage{
sp_tessellation_polygons(sp)
}
\arguments{
\item{sp}{Spatial points object.}
}
\value{
Spatial polygons or spatial polygons data frame depending on 
\code{sp}.
}
\description{
Tessellation polygons have no gaps and the Dirichlet tessellation method is
used.
}
\author{
Stuart K. Grange
}
