% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_unite.R
\name{sp_unite}
\alias{sp_unite}
\title{Function to unite spatial objects.}
\usage{
sp_unite(sp)
}
\arguments{
\item{sp}{Spatial object.}
}
\description{
\code{sp_unite} will join/combine/unite geometries which intersect.
}
\details{
This function is not complete.
}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{gUnion}}
}

