% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_wkt.R
\name{str_wkt}
\alias{str_wkt}
\title{Function to create WKT strings.}
\usage{
str_wkt(latitude, longitude, type = "point", round = NA)
}
\arguments{
\item{latitude}{Latitude vector.}

\item{longitude}{Longitude vector.}

\item{type}{Type of geom to create. Only \code{"points"} are supported 
currently.}

\item{round}{Decimal points to round \code{latitude} and \code{longitude}
vector.}
}
\value{
Character vector.
}
\description{
Function to create WKT strings.
}
\author{
Stuart K. Grange
}
