% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_drill.R
\name{tidy_ra_drill}
\alias{tidy_ra_drill}
\title{Function to reshape the return from \code{\link{ra_drill}} to be "tidy-data".}
\usage{
tidy_ra_drill(df)
}
\arguments{
\item{df}{Data frame/tibble from \code{\link{ra_drill}}}
}
\value{
Tibble.
}
\description{
Function to reshape the return from \code{\link{ra_drill}} to be "tidy-data".
}
\seealso{
\code{\link{ra_drill}}
}
\author{
Stuart K. Grange
}
