% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/projection_conversions.R
\name{wgs84_to_osgb36}
\alias{osgb36_to_wgs84}
\alias{wgs84_to_osgb36}
\title{Functions to convert points amoung different projection systems.}
\usage{
wgs84_to_osgb36(y, x)

osgb36_to_wgs84(x, y)
}
\description{
These functions use spatial libraries to do the coordinate conversion so the
sometimes many-step calculations do not have to be maintained.
}
\author{
Stuart K. Grange
}

