% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_geo_tiff.R
\name{write_geo_tiff}
\alias{write_geo_tiff}
\alias{write_esri_grid_ascii}
\title{Functions to export raster objects.}
\usage{
write_geo_tiff(ra, file)

write_esri_grid_ascii(ra, file, projection_file = FALSE)
}
\arguments{
\item{ra}{Raster object to export.}

\item{file}{File to export raster object too.}

\item{projection_file}{Should the projection string also be written?}
}
\value{
Invisible, raster object.
}
\description{
Functions to export raster objects.
}
\author{
Stuart K. Grange
}
