% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gml.R
\name{write_gml}
\alias{write_gml}
\title{Function to export geometries as a Geography Markup Language (GML) file.}
\usage{
write_gml(sp, file, layer = "spatial")
}
\arguments{
\item{sp}{Spatial object to be written as a GML/XML file.}

\item{file}{Filename of GML/XML file.}

\item{layer}{A string which represents the layer of the GML file where the 
geometries will be stored. Default uses a generic \code{spatial} string.}
}
\description{
\code{write_gml} uses \code{rgdal::writeOGR} for the spatial data writer
and will write projection information.
}
\examples{
\dontrun{

# Export polygons
write_gml(sp_zones, "~/Desktop/zone_polygons.gml", layer = "zones")

}

}
\author{
Stuart K. Grange
}

