% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gpx_with_dates.R
\name{write_gpx_with_dates}
\alias{write_gpx_with_dates}
\title{Function to export GPX files with a time variable.}
\usage{
write_gpx_with_dates(df, file)
}
\arguments{
\item{df}{Data frame or tibble with at least these three variables: 
\code{latitude}, \code{longitude}, and \code{date}. \code{date} also needs 
to be of \code{POSIXct} type.}

\item{file}{File name of GPX export.}
}
\value{
Invisible GPX markup.
}
\description{
Function to export GPX files with a time variable.
}
\author{
Stuart K. Grange
}
