% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_kml.R
\name{write_kml}
\alias{write_kml}
\title{Function to export geometries as a KML file.}
\usage{
write_kml(sp, file, force = TRUE)
}
\arguments{
\item{sp}{Spatial object.}

\item{file}{File name of KML file.}

\item{force}{Should the projection be forced to WGS84? Default is \code{TRUE}
because this is the standard for KML files.}
}
\value{
Invisible.
}
\description{
Function to export geometries as a KML file.
}
\examples{
\dontrun{

# Export a KML file containing points
write_kml(sp_bus_stations, "~/Desktop/bus_stations.kml")

# Export KML file which contains a track
write_kml(sp_track, "~/Desktop/drive_to_bath.kml")

}

}
\author{
Stuart K. Grange
}
