% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_to_spatial.R
\name{data_frame_to_points}
\alias{data_frame_to_points}
\title{Function for converting a data frame to a SpatialPointsDataFrame.}
\usage{
data_frame_to_points(df, latitude = "latitude", longitude = "longitude",
  projection = "+proj=longlat +datum=WGS84 +no_defs")
}
\arguments{
\item{df}{Data frame to be converted into SpatialPointsDataFrame.}

\item{latitude}{\code{df}'s latitude variable name.}

\item{longitude}{\code{df}'s longitude variable name.}

\item{projection}{\code{df}'s latitude and longitude projection system. 
Default is WGS84.}
}
\description{
Function for converting a data frame to a SpatialPointsDataFrame.
}
\examples{
\dontrun{
sp_points <- data_frame_to_points(data_monitoring_sites, "latitude", "longitude")
}

}
\author{
Stuart K. Grange
}

