% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_node_data.R
\name{get_osm_way_data}
\alias{get_osm_way_data}
\title{Function to scrape OpenStreetMap's way XML documents for data.}
\usage{
get_osm_way_data(id, progress = "none")
}
\arguments{
\item{id}{A vector os OpenStreetMap way IDs.}

\item{progress}{Type of progress bar to display.}
}
\value{
Named list or a list of named lists.
}
\description{
Function to scrape OpenStreetMap's way XML documents for data.
}
\examples{
\dontrun{

# Get data for York Minster
get_osm_way_data(28750371)

# Get data for two ways within a walking route (a relation)
get_osm_way_data(c(52461049, 233498222))

}

}
\author{
Stuart K. Grange
}
