% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sunrise.R
\name{get_sunrise}
\alias{get_sunrise}
\title{Function to calculate sunrise and sunset dates and times for a location.}
\usage{
get_sunrise(latitude, longitude, start = NA, end = NA, json = FALSE)
}
\arguments{
\item{latitude}{Latitude of a location.}

\item{longitude}{Longitude of a location.}

\item{start}{Start date.}

\item{end}{End date}

\item{json}{Should the return be a json string rather than a data frame? 
Useful for single observations.}
}
\value{
Tidy data frame or pretty printed JSON.
}
\description{
\code{get_sunrise} uses \code{\link{sunriset}} for the calculation of dates
and times.
}
\examples{

\dontrun{
# Get sunrise and sunset dates for London for the current year
get_sunrise(latitude = 51.5072, longitude = 0.1275)

# Or specify dates
get_sunrise(latitude = 51.5072, longitude = 0.1275, start = "2015-12-01", 
  end = "2015-12-31")
  
# Or as json
get_sunrise(latitude = 51.5072, longitude = 0.1275, json = TRUE)

}

}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{sunriset}}
}

