% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_elevation.R
\name{google_elevation}
\alias{google_elevation}
\title{Function to find elevation from latitude and longitude coordinate pairs using
Google Maps API.}
\usage{
google_elevation(latitude, longitude,
  key = "AIzaSyDoyFKU02pDdn-9k6DLFBcc1Rx_iwEZ9UE")
}
\arguments{
\item{latitude}{Numeric vector of latitude values.}

\item{longitude}{Numeric vector of longitude values.}

\item{key}{Google Maps API key.}
}
\description{
The Google Maps API is not open-source, therefore check the licensing 
conditions for usage conditions 
(\url{https://developers.google.com/maps/terms}). Users of this function 
must use the elevation data to display on a Google Map.
}
\details{
\code{google_elevation} returns a data frame with the latitude and longitude
pairs, elevation estimate, and (horizontal) resolution of elevation estimate. 
Coordinate pairs which contain \code{NA} will be returned as \code{NA} 
observation.
}
\author{
Stuart K. Grange
}

