% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_geocode.R
\name{google_geocode}
\alias{google_geocode}
\title{Function to find latitude and longitude pairs from an address-string using 
Google Maps API.}
\usage{
google_geocode(input, source = "google", wkt = FALSE, verbose = TRUE)
}
\arguments{
\item{input}{The input to transform to latitude and longitude pairs.}

\item{source}{What API should \code{google_geocode} access? Options are 
"google" or "dsk" and the default is "google".}

\item{wkt}{Should the return be a vector of WKT (well known text)? This is 
useful for when transforming a data frame.}

\item{verbose}{Should the functions give messages?}
}
\value{
Data frame or character vector.
}
\description{
The Google Maps API is not open-source, therefore check the licensing 
conditions for usage conditions 
(\url{https://developers.google.com/maps/terms}). Users of this function 
must use the geocoded data to display on a Google Map.
}
\details{
\code{google_geocode} is a wrapper for \code{ggmap::geocode} with some simple 
enhancements to clean the output and avoid messages to the console. Other 
address elements such as postcodes can also be transformed successfully.
}
\examples{
\dontrun{

# Return data frame
google_geocode("bath abbey")

# Only a vector without messages
google_geocode("york minster", verbose = FALSE, wkt = TRUE))

}

}
\author{
Stuart K. Grange
}
