% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_plot.R
\name{leaflet_plot}
\alias{leaflet_plot}
\title{Convenience function to quickly plot geometries on a leaflet map.}
\usage{
leaflet_plot(sp, popup = NULL, force = TRUE, colour = "#03F",
  opacity = 0.5, fill_opacity = 0.2)
}
\arguments{
\item{sp}{Spatial object to be plotted.}

\item{popup}{Vector of variables to be used as a popup on the map.}

\item{force}{Should the projection be forced to WGS84? Default is \code{TRUE}.}

\item{colour}{Colour of geometry.}

\item{opacity}{Opacity of the edge of the geometry.}

\item{fill_opacity}{Internal opacity of the geometry.}
}
\value{
Invisible, a leaflet map.
}
\description{
Convenience function to quickly plot geometries on a leaflet map.
}
\author{
Stuart K. Grange
}
