% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_plot.R
\name{leaflet_plot}
\alias{leaflet_plot}
\title{Convenience function to quickly plot geometries on a leaflet map.}
\usage{
leaflet_plot(sp, popup = NULL, force = TRUE)
}
\arguments{
\item{sp}{Spatial object to be plotted.}

\item{popup}{Popup variable to add to map.}

\item{force}{Should the projection be forced to WGS84? Default is \code{TRUE}.}
}
\description{
Convenience function to quickly plot geometries on a leaflet map.
}
\author{
Stuart K. Grange
}

