% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_sp.R
\name{map_to_sp}
\alias{map_to_sp}
\title{Function to convert map database objects to spatial data.}
\usage{
map_to_sp(map = "world")
}
\description{
Function to convert map database objects to spatial data.
}
\author{
Stuart K. Grange
}

