% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_sp.R
\name{map_to_sp}
\alias{map_to_sp}
\title{Function to convert map database objects to spatial data.}
\usage{
map_to_sp(map = "world")
}
\arguments{
\item{map}{Map database to turn into spatial data. \code{map} can be either
\code{"map"} or \code{"world_hires"}.}
}
\value{
SpatialPolygonsDataFrame with \code{id} and \code{name} variables.
}
\description{
Function to convert map database objects to spatial data.
}
\examples{
\dontrun{
# Lower resolution
sp_map <- map_to_sp(map = "world")

#
sp_map_better <- map_to_sp(map = "world_hires")

}

}
\author{
Stuart K. Grange
}

