% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_wgs84.R
\name{projection_wgs84}
\alias{projection_wgs84}
\alias{projection_google}
\alias{projection_bng}
\alias{projection_nztm}
\alias{projection_us_albers}
\alias{projection_antarctic}
\alias{projection_norway}
\alias{projection_lcc}
\alias{projection_swiss}
\alias{projection_slovenia}
\alias{projection_utm}
\title{Functions to wrap common projection (PROJ.4) strings.}
\usage{
projection_wgs84()

projection_google()

projection_bng()

projection_nztm()

projection_us_albers()

projection_antarctic()

projection_norway()

projection_lcc()

projection_swiss()

projection_slovenia()

projection_utm()
}
\description{
Functions to wrap common projection (PROJ.4) strings.
}
\author{
Stuart K.Grange
}
