% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_wgs84.R
\name{projection_wgs84}
\alias{projection_wgs84}
\alias{projection_google}
\alias{projection_bng}
\alias{projection_nztm}
\alias{projection_us_albers}
\alias{projection_antarctic}
\alias{projection_norway}
\alias{projection_lcc}
\alias{projection_swiss}
\alias{projection_slovenia}
\alias{projection_utm}
\alias{projection_madrid}
\alias{projection_mollweide}
\alias{projection_china}
\alias{projection_beijing}
\alias{projection_etr89}
\title{Functions to wrap common projection (PROJ.4) strings.}
\usage{
projection_wgs84()

projection_google()

projection_bng()

projection_nztm()

projection_us_albers()

projection_antarctic()

projection_norway()

projection_lcc()

projection_swiss()

projection_slovenia()

projection_utm()

projection_madrid()

projection_mollweide()

projection_china()

projection_beijing()

projection_etr89()
}
\value{
Character vector with length of 1.
}
\description{
Functions to wrap common projection (PROJ.4) strings.
}
\seealso{
\href{http://spatialreference.org/}{spatialreference.org}
}
\author{
Stuart K.Grange
}
