% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_bind.R
\name{ra_bind}
\alias{ra_bind}
\title{Function to bind multiple raster objects together.}
\usage{
ra_bind(..., tolerance = 0.05)
}
\arguments{
\item{...}{Raster objects.}

\item{tolerance}{Permissible difference in origin.}
}
\value{
Raster object.
}
\description{
Function to bind multiple raster objects together.
}
\seealso{
\code{\link[raster]{merge}}
}
\author{
Stuart K. Grange
}
