% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_mask.R
\name{ra_mask}
\alias{ra_mask}
\title{Function to mask/filter/crop a raster object to a spatial polygon.}
\usage{
ra_mask(ra, sp_polygon)
}
\arguments{
\item{ra}{Raster object to be masked/filtered.}

\item{sp_polygon}{Spatial polygon to be used as the filter.}
}
\description{
Function to mask/filter/crop a raster object to a spatial polygon.
}
\author{
Stuart K. Grange
}
