% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_read.R
\name{ra_read}
\alias{ra_read}
\title{Function to read a raster object from file.}
\usage{
ra_read(file, in_memory = FALSE)
}
\arguments{
\item{file}{File containing raster data.}

\item{in_memory}{Should the raster object be forced into physical memory?}
}
\value{
Raster object.
}
\description{
Function to read a raster object from file.
}
\seealso{
\code{\link[raster]{raster}}
}
\author{
Stuart K. Grange
}
