% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_gpx.R
\name{scrape_gpx}
\alias{scrape_gpx}
\title{Function to read gpx files as a data frame from the GPX Logger Android phone
app.}
\usage{
scrape_gpx(file, transform = TRUE)
}
\arguments{
\item{file}{GPX file.}

\item{transform}{Should latitude and longitude be used to calculate speed and
distance too?}
}
\value{
Data frame with correct data types.
}
\description{
Function to read gpx files as a data frame from the GPX Logger Android phone
app.
}
\author{
Stuart K. Grange
}

