% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_bind.R
\name{sp_bind}
\alias{sp_bind}
\title{Function to bind different spatial objects together.}
\usage{
sp_bind(...)
}
\arguments{
\item{...}{Spatial objects of the same type. \code{...} can be a number of 
individual objects or a list containing many objects.}
}
\description{
\code{sp_bind} combines spatial objects of the same type together. This 
function is useful for combining shapefiles or tables of WKT strings
together. The feature IDs within line- and polygon-spatial objects must be 
manipulated for the binding to occur. The original, non-manipulated IDs are 
not preserved. \code{sp_bind} will still bind objects if their data slots
contain different variables.
}
\details{
For uniting geometries which exist in the same data object, use 
\code{\link{sp_unite}}.
}
\examples{
\dontrun{

# Two objects
sp_london_two <- sp_bind(sp_croydon, sp_islington)

# Many objects
sp_london_many <- sp_bind(sp_croydon, sp_islington, sp_bexley, sp_brent)

# Or a list of many objects
sp_london_many_many <- sp_bind(sp_list_of_many)


# Load many files with plyr then bind together
# Get files
file_list <- list.files("data/shapefiles", "shp")

# Load files
sp_list <- llply(file_list, sp_read, verbose = FALSE)

# Bind all spatial objects
sp_bound <- sp_bind(sp_list)

}

}
\seealso{
\code{\link{sp_unite}}
}
\author{
Stuart K. Grange
}
