% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp_bind.R
\name{sp_bind}
\alias{sp_bind}
\alias{sp_bind_many}
\title{Functions to bind/combine spatial objects together.}
\usage{
sp_bind(sp, sp.2, force = TRUE)

sp_bind_many(sp.list, force = FALSE, progress = TRUE)
}
\arguments{
\item{sp}{Spatial object one.}

\item{sp.2}{Spatial object two.}

\item{force}{Should the feature IDs be forced to change? This may be
necessary when feature IDs are not sequential.}

\item{sp.list}{A list containing spatial two or more spatial objects.}
}
\description{
\code{sp_bind} combines two similar spatial objects together with the same
syntax of \code{rbind}. Currently, only two objects can be bound together.
\code{sp_bind_many} combines multiple spatial objects together but
the input must be a list.
}
\details{
These function are useful for combining shapefiles or vectors of WKT strings
together.

The feature IDs within line- and polygon-spatial objects must be manipulated
for the binding to occur. The original, non-manipulated IDs are not
preserved.

To-do: This function accumulates a spatial object in a for-loop which is
inefficient. The binding process should be handled by \code{do.call("rbind")}.
This needs to be solved soon.
}
\examples{
\dontrun{
# Two objects
sp.london.two <- sp_bind(sp.croydon, sp.islington)

# Many objects, note the nested list
sp.london.many <- sp_bind_many(list(sp.croydon, sp.islington, sp.bexley,
  sp.brent))
}
}
\author{
Stuart K. Grange
}

