% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_bind.R
\name{sp_bind}
\alias{sp_bind}
\alias{sp_bind_many}
\title{Functions to bind spatial objects together.}
\usage{
sp_bind(sp, sp_2)

sp_bind_many(sp_list)
}
\arguments{
\item{sp}{Spatial object one.}

\item{sp_2}{Spatial object two.}

\item{sp_list}{A list containing two or more spatial objects.}
}
\description{
\code{sp_bind} combines two similar spatial objects together with the same 
syntax of \code{rbind}. \code{sp_bind_many} combines multiple spatial objects 
together but the input must be a list.
}
\details{
These function are useful for combining shapefiles or tables of WKT strings
together. The feature IDs within line- and polygon-spatial objects must be 
manipulated for the binding to occur. The original, non-manipulated IDs are 
not preserved.
}
\examples{
\dontrun{
# Two objects
sp_london_two <- sp_bind(sp_croydon, sp_islington)

# Many objects, note the nested list
sp_london_many <- sp_bind_many(list(sp_croydon, sp_islington, sp_bexley, 
  sp_brent))
}

}
\author{
Stuart K. Grange
}

