% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_create_envelope.R
\name{sp_create_envelope}
\alias{sp_create_envelope}
\title{Function to create a rectangular polygon envelope, usually used for filtering.}
\usage{
sp_create_envelope(envelope,
  projection = "+proj=longlat +datum=WGS84 +no_defs")
}
\arguments{
\item{envelope}{A vector with the length of 4 containing the extent of the 
envelope to be used as a bounding box. The order is: xmin, xmax, ymin, ymax.
Watch out when using latitude and longitude, the order is longitude, 
longitude, then latitude, latitude.}

\item{projection}{A proj4 string for the projection-transformation. Default 
is the WGS84 string.}
}
\description{
Function to create a rectangular polygon envelope, usually used for filtering.
}
\author{
Stuart K. Grange
}

