% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_fix.R
\name{sp_fix}
\alias{sp_fix}
\title{Function to fix polygons when they have issues with validity and holes.}
\usage{
sp_fix(sp, print = TRUE)
}
\arguments{
\item{sp}{Spatial object. Polygons are currently the only object-type which
are supported.}

\item{print}{Should the features which are corrected, and why be displayed to
the user? Default is \code{TRUE}.}
}
\description{
Function to fix polygons when they have issues with validity and holes.
}
\examples{
\dontrun{
sp_clean <- sp_fix(sp_pm10)
}

}
\author{
Stuart K. Grange
}

