% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_from_js_array.R
\name{sp_from_js_array}
\alias{sp_from_js_array}
\title{Function to create spatial objects from JavaScript arrays.}
\usage{
sp_from_js_array(text, type = "points",
  projection = projection_wgs84())
}
\arguments{
\item{text}{Character vector, file, or url containing a JavaScript array.}

\item{type}{Spatial data type. Default is \code{"points"}.}

\item{projection}{Projection string.}
}
\value{
Spatial object.
}
\description{
JavaScript arrays take the form of:
\code{[[54.35221,-0.88518],[54.35237,-0.88544]]}.
}
\author{
Stuart K. Grange
}
