% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_intersect.R
\name{sp_intersect}
\alias{sp_intersect}
\title{Function to test weather one spatial object intersects with another.}
\usage{
sp_intersect(sp_one, sp_two, features = TRUE, simplify = FALSE, check = FALSE)
}
\arguments{
\item{sp_one}{First spatial object to be tested.}

\item{sp_two}{Second spatial object to be tested.}

\item{features}{Should the calculation be calculated on every feature of 
\code{sp_one} and \code{sp_two}.}

\item{simplify}{Should the return be simplified to a logical vector?}

\item{check}{Should the intersection function perform some checks?}
}
\value{
Logical matrix or vector.
}
\description{
Function to test weather one spatial object intersects with another.
}
\seealso{
\code{\link{sp_filter}}, \code{\link{sp_crop}}
}
\author{
Stuart K. Grange
}
