% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_move.R
\name{sp_move}
\alias{sp_move}
\alias{sp_shift}
\title{Function to move/shift a spatial object.}
\usage{
sp_move(sp, x, y)

sp_shift(sp, x, y)
}
\arguments{
\item{sp}{Spatial object.}

\item{x}{Offset amount (in \code{sp} projection's units) in left-right 
dimension.}

\item{y}{Offset amount (in \code{sp} projection's units) in up-down 
dimension.}
}
\value{
Spatial object.
}
\description{
Function to move/shift a spatial object.
}
\seealso{
\code{\link{elide}}
}
\author{
Stuart K. Grange
}
