% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp_read.R
\name{sp_read}
\alias{sp_read}
\title{Function to read spatial data files in a consistent way.}
\usage{
sp_read(file, verbose = TRUE, tolower = TRUE)
}
\arguments{
\item{file}{Spatial data file.}

\item{verbose}{Should information about the data be printed when being
loaded? Default is TRUE.}

\item{tolower}{Should the names of the data slot be forced to be lower case?}
}
\description{
\code{sp_read} is a wrapper for \code{rgdal::readOGR}, but its usage is the
same as the other file readers in R. Unlike \code{rgdal::readOGR},
\code{sp_read} uses a single file-string which does not have to be expanded
and can handle shapefiles with or without specified extensions.
}
\details{
GPX files containing tracks, routes, or waypoints can be read with
\code{sp_read} as can GeoJSON files.
}
\examples{
\dontrun{

# Load a shapefile without extension
sp_london <- sp_read("london_borough")

# Load a shapefile with an extension within a home area
sp_london <- sp_read("~/Desktop/london_borough.shp")

# Load a gpx file, no need to define layer
sp_hira <- sp_read("~/Desktop/hira_mtb_park.gpx")

# Load GeoJSON file
sp_thames_locks <- sp_read("thames_locks.json", verbose = FALSE)

}
}
\author{
Stuart K. Grange
}

