% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp_read.R
\name{sp_read}
\alias{sp_read}
\title{Function to read spatial data files in a consistent way.}
\usage{
sp_read(file, verbose = FALSE)
}
\arguments{
\item{file}{Spatial data file.}

\item{verbose}{Should information about the data be printed when being
loaded?}
}
\description{
\code{sp_read} is a wrapper for \code{rgdal::readOGR}, but its usage is the
same as the other file readers in R. Unlike \code{rgdal::readOGR},
\code{sp_read} uses a single file string which does not have to be expanded
and can handle shape files with or without specified extensions. GPX files
can also be read with \code{sp_read}.
}
\examples{
\dontrun{
# Load a shape file without extension
sp.london <- sp_read("london_borough")

# Load a shape file with an extension within a home area
sp.london <- sp_read("~/Desktop/london_borough.shp")

# Load a gpx file
sp.hira <- sp_read("~/Desktop/hira_mtb_park.gpx")
}
}
\author{
Stuart K. Grange
}

