% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_sample_n.R
\name{sp_sample_n}
\alias{sp_sample_n}
\title{Function to randomly sample n features in a spatial object.}
\usage{
sp_sample_n(sp, n)
}
\arguments{
\item{sp}{Spatial object.}

\item{n}{Number of features to sample.}
}
\description{
Function to randomly sample n features in a spatial object.
}
\seealso{
\code{\link{set.seed}}
}
\author{
Stuart K. Grange
}
