% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_segment_lines.R
\name{sp_segment_lines}
\alias{sp_segment_lines}
\title{Function to split spatial lines into segments with fixed lengths.}
\usage{
sp_segment_lines(sp, length, features = FALSE)
}
\arguments{
\item{sp}{Spatial line object.}

\item{length}{Length for the split segments. This is in the units of 
\code{sp}'s projection system.}

\item{features}{Should the function be applied to all features in \code{sp}
separately?}
}
\value{
SpatialLinesDataFrame with one \code{id} variable.
}
\description{
\code{sp_segment_lines} has been pulled from another package which can be
found \href{https://github.com/jmt2080ad/polylineSplitter}{here},
}
\seealso{
\url{https://github.com/jmt2080ad/polylineSplitter}
}
\author{
Jason Taylor, David Blodgett, and Stuart K. Grange
}
