% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_simplify.R
\name{sp_simplify}
\alias{sp_simplify}
\title{Function to simplify spatial objects using the Douglas-Peuker algorithm.}
\usage{
sp_simplify(sp, tolerance, preserve = TRUE)
}
\arguments{
\item{sp}{Spatial object which is to be simplified.}

\item{tolerance}{Tolerance value to be used by the Douglas-Peuker algorithm.}

\item{preserve}{Should the algorithm attempt to preserve the topology? 
Default is \code{TRUE}.}
}
\description{
Holes will appear between the geometries at times.
}
\examples{
\dontrun{

sp_simple <- sp_simplify(sp, 0.01)

}

}
\author{
Stuart K. Grange
}

