% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_snap_points_to_lines.R
\name{sp_snap_points_to_lines}
\alias{sp_snap_points_to_lines}
\title{Function to "snap" spatial points to the nearest spatial line.}
\usage{
sp_snap_points_to_lines(
  sp_points,
  sp_lines,
  max_distance = NA,
  id = NA,
  rename = TRUE
)
}
\arguments{
\item{sp_points}{Spatial points to be snapped.}

\item{sp_lines}{Spatial lines to snap \code{sp_points} to.}

\item{max_distance}{The maximum distance to avoid snapping points that are 
farther apart.}

\item{id}{The variable in \code{sp_lines} which will be transferred to 
\code{sp_points} to distinguish the line which each point was snapped to.}

\item{rename}{Should the \code{id} variable be renamed?}
}
\value{
Spatial points with altered coordinates; all points will lie on lines
supplied for \code{sp_lines}.
}
\description{
Function to "snap" spatial points to the nearest spatial line.
}
\seealso{
\link{snapPointsToLines}
}
\author{
Stuart K. Grange
}
