% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_unite.R
\name{sp_unite}
\alias{sp_unite}
\title{Function to unite spatial objects.}
\usage{
sp_unite(sp)
}
\arguments{
\item{sp}{Spatial object.}
}
\description{
\code{sp_unite} will join/combine/unite geometries which are contained in the
same data object. For example, use \code{sp_unite} when a data object contains
many individual polygons but they form the same unit and therefore they are 
considered sub-geometries.
}
\details{
For binding different spatial objects together, use \code{\link{sp_bind}}.
}
\seealso{
\code{\link{gUnion}}, \code{\link{sp_bind}}
}
\author{
Stuart K. Grange
}
