% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_geojson.R
\name{write_geojson}
\alias{write_geojson}
\title{Function to write GeoJSON files with usage analogous to \code{write.table}.}
\usage{
write_geojson(sp, file, pretty = TRUE)
}
\arguments{
\item{sp}{Spatial object to be written as GeoJSON file.}

\item{file}{File name of GeoJSON file.}

\item{pretty}{Format the JSON for readability.}
}
\description{
Function to write GeoJSON files with usage analogous to \code{write.table}.
}
\examples{
\dontrun{
# Export spatial object as GeoJSON file
write_geojson(sp_thames_locks, "thames_locks.json")

}
}
\author{
Stuart K. Grange
}

